% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{dracula_button}
\alias{dracula_button}
\alias{update_dracula_button}
\title{Dracula action button}
\usage{
dracula_button(
  inputId,
  label,
  icon = NULL,
  width = NULL,
  ...,
  color = "purple",
  textColor = "black",
  disabled = FALSE,
  outline = FALSE,
  ghost = FALSE,
  size = "md"
)

update_dracula_button(
  inputId,
  label = NULL,
  icon = NULL,
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{The contents of the button or link--usually a text label, but
you could also use any other HTML, like an image.}

\item{icon}{An optional \code{\link[shiny:icon]{icon()}} to appear on the button.}

\item{width}{The width of the input, e.g. \code{'400px'}, or \code{'100\%'};
see \code{\link[shiny:validateCssUnit]{validateCssUnit()}}.}

\item{...}{Named attributes to be applied to the button or link.}

\item{color}{Input color.}

\item{textColor}{Text color.}

\item{disabled}{Whether to disable the checkbox.
Default to FALSE.}

\item{ghost}{Ghost style.}

\item{session}{The \code{session} object passed to function given to
\code{shinyServer}. Default is \code{getDefaultReactiveDomain()}.}
}
\value{
Send message to JavaScript
}
\description{
Dracula action button

Update dracula action button
}
\note{
icon does not work yet.
}
\examples{
if (interactive()) {
 library(shiny)
 library(dRacula)
 ui <- dracula_page(
   dracula_button(
    "btn",
    "My action button",
    color = "animated"
   )
 )
 server <- function(input, output, session) {
   observe({
    print(input$btn)
   })
 }
 shinyApp(ui, server)

 # Update button #
 ui <- dracula_page(
  dracula_button("go", "Update text"),
  dracula_button("button", "My button", color = "animated")
 )
 server <- function(input, output, session) {
   observeEvent(input$go, {
     update_dracula_button(
       "button",
       label = "New label",
       icon = icon("folder")
     )
   })
 }
 shinyApp(ui, server)
}
}
\seealso{
\url{https://ui.draculatheme.com/button} for all
available options.
}
