% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dracula_ui-dependencies.R
\name{add_dracula_ui_deps}
\alias{add_dracula_ui_deps}
\title{dracula-ui dependencies utils}
\usage{
add_dracula_ui_deps(tag)
}
\arguments{
\item{tag}{Element to attach the dependencies.}
}
\description{
This function attaches dracula-ui. dependencies to the given tag
}
