% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dRacula-package.R
\docType{package}
\name{dRacula-package}
\alias{dRacula}
\alias{dRacula-package}
\title{dRacula: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: David Granjon \email{dgranjon@ymail.com}

}
\keyword{internal}
