% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{dracula_avatar}
\alias{dracula_avatar}
\title{Dracula avatar}
\usage{
dracula_avatar(src, color = "purple", borderSize = "md")
}
\arguments{
\item{src}{Background image path.}

\item{color}{Background color.}

\item{borderSize}{Border size.}
}
\description{
Dracula avatar
}
\examples{
if (interactive()) {
 library(shiny)
 library(dRacula)
 ui <- dracula_page(
  dracula_avatar(
    "https://ui.draculatheme.com/static/images/avatar.png",
    color = "green"
  ),
  dracula_avatar(
    "https://ui.draculatheme.com/static/images/avatar.png",
    borderSize = "lg"
  )
 )
 server <- function(input, output, session) {
 }
 shinyApp(ui, server)
}
}
\seealso{
See \url{https://ui.draculatheme.com/avatar#properties}.
}
