% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{dracula_badge}
\alias{dracula_badge}
\title{Dracula badge}
\usage{
dracula_badge(..., color, outline = FALSE, size = "md")
}
\arguments{
\item{...}{Card content.}

\item{color}{Background color.}

\item{outline}{Alternate appearance.}

\item{size}{Padding and margin.}
}
\description{
Dracula badge
}
\examples{
if (interactive()) {
 library(shiny)
 library(dRacula)
 ui <- dracula_page(
  dracula_badge("Hello World", color = "animated"),
  dracula_badge(
    "1",
    color = "red",
    outline = TRUE
  )
 )
 server <- function(input, output, session) {
 }
 shinyApp(ui, server)
}
}
