% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dracula_palette.R
\name{dracula_bright_palette}
\alias{dracula_bright_palette}
\title{Dracula Bright Palette}
\usage{
dracula_bright_palette()
}
\description{
Used in conjunction with \code{dracula_discrete_bright_palette} as an
internal closure for interfacing with \code{scale_fill_dracula} and \code{scale_color_dracula}.
}
