% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{dracula_card}
\alias{dracula_card}
\title{Dracula card container}
\usage{
dracula_card(
  ...,
  color = "purple",
  glowColor = NULL,
  borderColor = "purple",
  size = "md",
  rounded = "full",
  outline = FALSE,
  height = "auto",
  width = "auto",
  scrollbar = "purple"
)
}
\arguments{
\item{...}{Card content.}

\item{color}{Background color.}

\item{glowColor}{Glow color.}

\item{borderColor}{Border color. Only if
outline is TRUE.}

\item{size}{Padding and margin.}

\item{rounded}{Rounded corners.}

\item{outline}{Alternate appearance.}

\item{height}{Card height.}

\item{width}{Card width.}

\item{scrollbar}{Scrollbar color.}
}
\description{
Dracula card container
}
\examples{
if (interactive()) {
 library(shiny)
 library(dRacula)
 ui <- dracula_page(
   dracula_card(
    "My first card"
   ),
  dracula_card(
    "My second card",
    color = "pink",
    borderColor = "pink",
    outline = TRUE
  )
 )
 server <- function(input, output, session) {
 }
 shinyApp(ui, server)
}
}
\seealso{
\url{https://ui.draculatheme.com/card}
}
