% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dracula_palette.R
\name{dracula_discrete_bright_palette}
\alias{dracula_discrete_bright_palette}
\title{Dracula Discrete Bright Palette}
\usage{
dracula_discrete_bright_palette(n)
}
\arguments{
\item{n}{Number of colors to return.
If missing, defaults to the length of the entire palette}
}
\value{
Function for interfacing with scale functions
}
\description{
Used in conjunction with \code{dracula_bright_palette} as an
internal closure for interfacing with \code{scale_fill_dracula} and \code{scale_color_dracula}
}
