% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typography.R
\name{dracula_heading}
\alias{dracula_heading}
\title{Dracula heading tags}
\usage{
dracula_heading(..., tagName = "h1", color = "purple", size = "xl")
}
\arguments{
\item{...}{Text.}

\item{tagName}{Tag function: h1 until h6.}

\item{color}{Text color.}

\item{size}{Text size.}
}
\description{
Dracula heading tags
}
\examples{
if (interactive()) {
 library(shiny)
 library(dRacula)
 ui <- dracula_page(
  dracula_heading("Hello World", "h2"),
  dracula_heading(
    "Hello World",
    "h2",
    color = "pink-purple",
    size = "2xl"
  )
 )
 server <- function(input, output, session) {
 }
 shinyApp(ui, server)
}
}
\seealso{
\url{https://ui.draculatheme.com/heading}.
}
