% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{dracula_input}
\alias{dracula_input}
\title{Dracula input builder}
\usage{
dracula_input(
  inputId,
  label,
  value = NULL,
  placeholder = NULL,
  type = c("checkbox", "number", "text", "date", "range", "radio"),
  color = dracula_basic_colors,
  textColor = dracula_basic_colors,
  size = c("sm", "md", "lg"),
  outline = FALSE,
  borderSize = c("sm", "md", "lg")
)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{value}{Initial value.}

\item{placeholder}{A character string giving the user a hint as to what can
be entered into the control. Internet Explorer 8 and 9 do not support this
option.}

\item{type}{Input type.}

\item{color}{Input color.}

\item{textColor}{Text color.}

\item{size}{Input size.}

\item{outline}{Outline variant.}

\item{borderSize}{Border size.}
}
\description{
Internal builder for dracula inputs
}
\keyword{internal}
