% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typography.R
\name{dracula_link}
\alias{dracula_link}
\title{Dracula link}
\usage{
dracula_link(
  ...,
  href = "#",
  color = "white",
  weight = "normal",
  size = "md",
  hoverColor = "purple"
)
}
\arguments{
\item{...}{Link text.}

\item{href}{Link adress.}

\item{color}{Text color.}

\item{weight}{Text weight.}

\item{size}{Text size.}

\item{hoverColor}{Hover color.}
}
\description{
Dracula link
}
\examples{
if (interactive()) {
 library(shiny)
 library(dRacula)
 ui <- dracula_page(
  dracula_link(
    "Dracula UI",
    href = "https://ui.draculatheme.com/welcome",
    weight = "bold"
  )
 )
 server <- function(input, output, session) {
 }
 shinyApp(ui, server)
}
}
\seealso{
\url{https://ui.draculatheme.com/anchor#}.
}
