% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typography.R
\name{dracula_p}
\alias{dracula_p}
\title{Dracula paragraph}
\usage{
dracula_p(
  ...,
  color = "white",
  align = "left",
  weight = "normal",
  size = "md",
  lineHeight = "md"
)
}
\arguments{
\item{...}{Text.}

\item{color}{Text color.}

\item{align}{Text alignement.}

\item{weight}{Text weight.}

\item{size}{Text size.}

\item{lineHeight}{Text line height.}
}
\description{
Dracula paragraph
}
\examples{
if (interactive()) {
 library(shiny)
 library(dRacula)
 ui <- dracula_page(
  dracula_p(
    "Dracula is an 1897 Gothic horror novel by Irish author Bram Stoker. It
     introduced the character of Count Dracula and established many conventions
     of subsequent vampire fantasy. The novel tells the story of Dracula's
     attempt to move from Transylvania to England so that he may find new blood
     and spread the undead curse, and of the battle between Dracula and Professor
     Abraham Van Helsing.",
    color = "pink-purple",
    align = "center",
    weight = "bold",
    size = "lg",
    lineHeight = "xs"
  )
 )
 server <- function(input, output, session) {
 }
 shinyApp(ui, server)
}
}
\seealso{
\url{https://ui.draculatheme.com/paragraph}.
}
