% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{dracula_page}
\alias{dracula_page}
\title{Dracula page wrapper}
\usage{
dracula_page(...)
}
\arguments{
\item{...}{Shiny UI elements.}
}
\value{
An HTML tag
}
\description{
Provides necessary dependencies as well
as dark theme.
}
