% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggDracula.R
\name{dracula_palette}
\alias{dracula_palette}
\title{Dracula Palette}
\usage{
dracula_palette(num_cols, var_type = c("discrete", "continuous"))
}
\arguments{
\item{num_cols}{Number of colors to return}

\item{var_type}{Variable type}
}
\value{
Dracula palette
}
\description{
Dracula Palette
}
\examples{
dracula_palette(num_col  = nlevels(factor(mpg$manufacturer)), var_type = "discrete")
dracula_palette(var_type = "continuous")
}
