% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{dracula_range}
\alias{dracula_range}
\title{Dracula range input}
\usage{
dracula_range(
  inputId,
  label,
  value,
  min,
  max,
  step = NA,
  color = "white",
  textColor = "white",
  size = "md",
  borderSize = "lg"
)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{value}{Initial value.}

\item{min}{Minimum allowed value}

\item{max}{Maximum allowed value}

\item{step}{Interval to use when stepping between min and max}
}
\description{
Dracula range input
}
\examples{
if (interactive()) {
 library(shiny)
 library(dRacula)
 ui <- dracula_page(
   dracula_range(
    "range",
    "My range input",
    200,
    min = 10,
    max = 1000,
    step = 1,
    color = "orange"
   )
 )
 server <- function(input, output, session) {
   observe({
    print(input$range)
   })
 }
 shinyApp(ui, server)
}
}
