% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{dracula_switch}
\alias{dracula_switch}
\title{Dracula switch}
\usage{
dracula_switch(
  inputId,
  label,
  value = FALSE,
  color = "purple",
  textColor = "white",
  disabled = FALSE
)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{value}{Initial value (\code{TRUE} or \code{FALSE}).}

\item{color}{Input color.}

\item{textColor}{Text color.}

\item{disabled}{Whether to disable the checkbox.
Default to FALSE.}
}
\description{
Similar to checkbox but with
a toggle widget.
}
\examples{
if (interactive()) {
 library(shiny)
 library(dRacula)
 ui <- dracula_page(
   dracula_switch("switch", "My switch input", color = "pink")
 )
 server <- function(input, output, session) {
   observe({
    print(input$switch)
   })
 }
 shinyApp(ui, server)
}
}
