% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typography.R
\name{dracula_text}
\alias{dracula_text}
\title{Dracula text}
\usage{
dracula_text(
  ...,
  color = "white",
  align = "left",
  weight = "normal",
  size = "md"
)
}
\arguments{
\item{...}{Text.}

\item{color}{Text color.}

\item{align}{Text alignement.}

\item{weight}{Text weight.}

\item{size}{Text size.}
}
\description{
Dracula text
}
\examples{
if (interactive()) {
 library(shiny)
 library(dRacula)
 ui <- dracula_page(
  dracula_text("The quick vampire", color = "red")
 )
 server <- function(input, output, session) {
 }
 shinyApp(ui, server)
}
}
\seealso{
\url{https://ui.draculatheme.com/text}.
}
