% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typography.R
\name{dracula_text_cl}
\alias{dracula_text_cl}
\title{Generate class for text and p elements}
\usage{
dracula_text_cl(
  color = "white",
  align = "left",
  weight = "normal",
  size = "md"
)
}
\arguments{
\item{color}{Text color.}

\item{align}{Text alignement.}

\item{weight}{Text weight.}

\item{size}{Text size.}
}
\value{
String
}
\description{
Generate class for text and p elements
}
\keyword{internal}
