% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dracula_palette.R
\docType{data}
\name{dracula_tibble}
\alias{dracula_tibble}
\title{Dracula Tibble}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7 rows and 4 columns.
}
\usage{
dracula_tibble
}
\value{
Dracula Tibble
}
\description{
A Tibble of Dracula data that includes the palette specification.
See https://spec.draculatheme.com for details.
}
\keyword{datasets}
