% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggDracula-package.R
\docType{package}
\name{ggDracula-package}
\alias{ggDracula}
\alias{ggDracula-package}
\title{ggDracula: Dracula Palettes and ggplot2 Theme}
\description{
Provides a Dracula theme for ggplot2 and Dracula style palette.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dracula}
  \item \url{https://draculatheme.com}
  \item Report bugs at \url{https://github.com/dracula/ggplot2}
}

}
\author{
\strong{Maintainer}: Jordan Bradford \email{jrdnbradford@gmail.com}

}
\keyword{internal}
