% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_dracula}
\alias{scale_color_dracula}
\title{Dracula Color Scales for ggplot2}
\usage{
scale_color_dracula(..., discrete = FALSE, aesthetics = "color")
}
\arguments{
\item{...}{Arguments passed on to \code{continuous_scale} in \code{scale_colour_gradientn} or to \code{discrete_scale}}

\item{discrete}{Whether the scale is discrete. Defaults to \code{FALSE}}

\item{aesthetics}{The aesthetics for the plot. Defaults to \code{color}}
}
\description{
Dracula Color Scales for ggplot2
}
