% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_dracula}
\alias{scale_fill_dracula}
\title{Dracula Fill Scales for ggplot2}
\usage{
scale_fill_dracula(..., discrete = FALSE, aesthetics = "fill")
}
\arguments{
\item{...}{Arguments passed on to \code{continuous_scale} in \code{scale_fill_gradientn} or to \code{discrete_scale}}

\item{discrete}{Whether the scale is discrete. Defaults to \code{FALSE}}

\item{aesthetics}{The aesthetics for the plot. Defaults to \code{fill}}
}
\description{
Dracula Fill Scales for ggplot2
}
