% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{dracula_switch}
\alias{dracula_switch}
\alias{update_dracula_switch}
\title{Dracula switch}
\usage{
dracula_switch(
  inputId,
  label,
  value = FALSE,
  color = "purple",
  textColor = "white",
  disabled = FALSE
)

update_dracula_switch(
  inputId,
  label = NULL,
  value = NULL,
  placeholder = NULL,
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{value}{Initial value (\code{TRUE} or \code{FALSE}).}

\item{color}{Input color.}

\item{textColor}{Text color.}

\item{disabled}{Whether to disable the checkbox.
Default to FALSE.}

\item{session}{The \code{session} object passed to function given to
\code{shinyServer}. Default is \code{getDefaultReactiveDomain()}.}
}
\value{
Send message to JavaScript
}
\description{
Similar to checkbox but with
a toggle widget.
}
\note{
label does not work yet.
}
\examples{
if (interactive()) {
 library(shiny)
 library(dRacula)
 ui <- dracula_page(
   dracula_switch("switch", "My switch input", color = "pink")
 )
 server <- function(input, output, session) {
   observe({
    print(input$switch)
   })
 }
 shinyApp(ui, server)

 # Update switch #
 ui <- dracula_page(
  dracula_button("go", "Update text"),
  dracula_switch("checkbox", "My checkbox", color = "pink")
 )
 server <- function(input, output, session) {
   observeEvent(input$go, {
     update_dracula_switch(
       "checkbox",
       label = "Plop",
       value = "New value"
     )
   })
 }
 shinyApp(ui, server)
}
}
\seealso{
\url{https://ui.draculatheme.com/input} for all
available options.
}
