% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabs.R
\name{dracula_tabs}
\alias{dracula_tabs}
\alias{dracula_tab}
\title{Dracula tabset}
\usage{
dracula_tabs(..., id = NULL, selected = NULL, color = "white")

dracula_tab(title, ..., value = title, icon = NULL)
}
\arguments{
\item{...}{\code{\link[shiny:tabPanel]{tabPanel()}} elements to include in the tabset}

\item{id}{If provided, you can use \verb{input$}\emph{\code{id}} in your
server logic to determine which of the current tabs is active. The value
will correspond to the \code{value} argument that is passed to
\code{\link[shiny:tabPanel]{tabPanel()}}.}

\item{selected}{The \code{value} (or, if none was supplied, the \code{title})
of the tab that should be selected by default. If \code{NULL}, the first
tab will be selected.}

\item{color}{Tab color.}

\item{title}{Display title for tab}

\item{value}{The value that should be sent when \code{tabsetPanel} reports
that this tab is selected. If omitted and \code{tabsetPanel} has an
\code{id}, then the title will be used.}

\item{icon}{Optional icon to appear on the tab. This attribute is only
valid when using a \code{tabPanel} within a \code{\link[shiny:navbarPage]{navbarPage()}}.}
}
\description{
Dracula tabset

Dracula tab panel
}
\seealso{
\url{https://ui.draculatheme.com/tabs#}.
}
