% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_dracula.R
\name{theme_dracula}
\alias{theme_dracula}
\title{Theme Dracula}
\usage{
theme_dracula()
}
\description{
Provides a minimal \code{ggplot2} theme with a Dracula-ish, dark backdrop.
}
\examples{
library(ggplot2)
library(dplyr)

mpg \%>\%
  filter(manufacturer \%in\% c("honda", "ford", "dodge", "audi")) \%>\%
  group_by(manufacturer) \%>\%
  summarize(mean_hwy = mean(hwy)) \%>\%
  ggplot(aes(x = manufacturer, y = mean_hwy, fill = manufacturer)) +
  theme(legend.position = "none") +
  coord_flip() + geom_col() +
  scale_fill_dracula(discrete = TRUE) +
  theme_dracula()
}
