% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dRacula-dependencies.R
\name{add_dRacula_deps}
\alias{add_dRacula_deps}
\title{dRacula dependencies utils}
\usage{
add_dRacula_deps(tag)
}
\arguments{
\item{tag}{Element to attach the dependencies.}
}
\description{
This function attaches dRacula dependencies to the given tag
}
