% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dRacula-package.R
\docType{package}
\name{dRacula-package}
\alias{dRacula}
\alias{dRacula-package}
\title{dRacula: Dracula theme for R}
\description{
All things Dracula theme for R
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dracula/dRacula}
  \item \url{https://github.com/dracula}
  \item \url{https://draculatheme.com}
  \item \url{https://draculatheme.com/dRacula}
  \item Report bugs at \url{https://github.com/dracula/dRacula/issues}
}

}
\author{
\strong{Maintainer}: Jordan Bradford \email{jrdnbradford@gmail.com}

}
\keyword{internal}
