% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggDracula.R
\name{theme_dracula}
\alias{theme_dracula}
\title{Theme Dracula}
\usage{
theme_dracula()
}
\description{
Provides a minimal \code{ggplot2} theme with a Dracula-ish, dark backdrop.
}
\examples{
theme_set(theme_dracula())
}
